_Backend/requirements.txt
Flask>=3.0.0
flask-restx>=1.3.0
python-dotenv==1.0.0
Flask-Cors>=3.0.10
# Optional if Python<3.9:
# pytz>=2023.3
# Optional for deployment
# gunicorn>=21.2.0
Flask-SQLAlchemy==3.1.1